#17 refactor : DetailSignup BindingAdapter 별도 파일 분리
