+++ Aos/app/src/main/java/com/avengers/nibobnebob/presentation/ui/intro/signup/DetailSignupFragment.kt
@@ -85,34 +85,3 @@ class DetailSignupFragment :
     }
 }
 
-@BindingAdapter("helperMessage")
-fun bindHelpMessage(view: TextView, inputState: InputState) {
-    when (inputState) {
-        is InputState.Success -> {
-            view.text = inputState.msg
-            view.setTextColor(ContextCompat.getColor(view.context, R.color.nn_primary6))
-        }
-
-        is InputState.Error -> {
-            view.text = inputState.msg
-            view.setTextColor(ContextCompat.getColor(view.context, R.color.nn_rainbow_red))
-        }
-
-        else -> {
-            view.text = ""
-        }
-    }
-}
-
-@BindingAdapter("textLayoutColor")
-fun bindTextLayoutColor(view: TextInputLayout, inputState: InputState) {
-    when (inputState) {
-        is InputState.Success -> view.boxStrokeColor =
-            ContextCompat.getColor(view.context, R.color.nn_primary6)
-
-        is InputState.Error -> view.boxStrokeColor =
-            ContextCompat.getColor(view.context, R.color.nn_rainbow_red)
-
-        else -> view.boxStrokeColor = ContextCompat.getColor(view.context, R.color.nn_primary6)
-    }
-}
